requirements.txt
gunicorn==21.2.0
flask==3.1.0
python-dotenv==1.1.0
pypdf==5.6.1

langchain==0.3.26
langchain-community==0.3.26
langchain-pinecone==0.2.8
langchain-groq==0.3.8

pinecone-client==3.2.2
fastembed==0.3.6
